moved fit modals code to fit_modal.py
