+++ app/sims/fit/fit_modal.py
@@ -3,6 +3,7 @@ import dash_bootstrap_components as dbc
 import dash_core_components as dcc
 import dash_html_components as html
 from dash.dependencies import Input
+from dash.dependencies import MATCH
 from dash.dependencies import Output
 from dash.dependencies import State
 
@@ -77,3 +78,53 @@ def info_body():
 
 
 fit_info_modal = info_modal_ui()
+
+
+def make_modals(params_dict):
+    """Constructs hidden html.Div containing params modals
+
+    Params:
+        params_dict: dict representation of Parameters object
+
+    Returns:
+        list of modals
+    """
+
+    def make_modal(key, vals):
+        """Helper method to make each modal"""
+        min_id = {"name": f"{key}-min", "kind": "min"}
+        max_id = {"name": f"{key}-max", "kind": "max"}
+        expr_id = {"name": f"{key}-expr", "kind": "expr"}
+        modal_id = {"kind": "modal", "parrent": key}
+
+        min_ = html.Div(
+            ["Minimum", dcc.Input(value=vals["min"], id=min_id)], className="input-form"
+        )
+        max_ = html.Div(
+            ["Maximum", dcc.Input(value=vals["max"], id=max_id)], className="input-form"
+        )
+        expr = html.Div(
+            ["Expression", dcc.Input(value=vals["expr"], type="text", id=expr_id)],
+            className="input-form",
+        )
+
+        head = dbc.ModalHeader(html.B(key))
+        body = dbc.ModalBody([min_, max_, expr])
+
+        return dbc.Modal([head, body], id=modal_id)
+
+    modals = []
+    for key, vals in params_dict.items():
+        modals += [make_modal(key, vals)]
+
+    return modals
+
+
+# Opens/closes params modal
+app.clientside_callback(
+    "function (n1, is_open) { if(n1 == null) { return false; } return !is_open; }",
+    Output({"kind": "modal", "parrent": MATCH}, "is_open"),
+    Input({"kind": "modal-btn", "parrent": MATCH}, "n_clicks"),
+    State({"kind": "modal", "parrent": MATCH}, "is_open"),
+    prevent_initial_call=True,
+)
